Add 2 missing members to Exception (#11442)
*     Add 2 missing members to Exception
