+++ src/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -53,6 +53,7 @@
     <Compile Include="System\IntPtrTests.cs" />
     <Compile Include="System\IO\DirectoryNotFoundExceptionTests.cs" />
     <Compile Include="System\IO\DirectoryNotFoundException.InteropTests.cs" />
+    <Compile Include="System\IO\EndOfStreamExceptionTests.cs" />    
     <Compile Include="System\IO\Exceptions.HResults.cs" />
     <Compile Include="System\IO\Exceptions.Utility.cs" />
     <Compile Include="System\IO\FileLoadExceptionTests.cs" />
@@ -66,7 +67,7 @@
     <Compile Include="System\MissingMethodExceptionTests.cs" />
     <Compile Include="System\MulticastDelegateTests.cs" />
     <Compile Include="System\NullableTests.cs" />
-    <Compile Include="System\ExceptionTests.cs" />
+    <Compile Include="System\NullReferenceExceptionTests.cs" />
     <Compile Include="System\ObjectTests.cs" />
     <Compile Include="System\Runtime\CompilerServices\ConditionalWeakTableTests.cs" />
     <Compile Include="System\Runtime\CompilerServices\StrongBoxTests.cs" />
@@ -108,6 +109,7 @@
     <Compile Include="System\ExternalExceptionTests.cs" />
     <Compile Include="System\ApplicationExceptionTests.cs" />
     <Compile Include="System\EntryPointNotFoundExceptionTests.cs" />
+    <Compile Include="System\ExceptionTests.cs" />
     <Compile Include="System\ExecutionEngineExceptionTests.cs" />
     <Compile Include="System\NotFiniteNumberExceptionTests.cs" />
     <Compile Include="System\StackOverflowExceptionTests.cs" />
