+++ src/System.Runtime/tests/System/ExceptionTests.cs
@@ -4,69 +4,37 @@
 
 using System;
 using System.IO;
+using System.Reflection;
 using Xunit;
 
 namespace System.Tests
 {
-    public static class NullReferenceExceptionTests
+    public static class ExceptionTests
     {
-        private const string exceptionMessage = "Created an exception";
-        private const string innerExceptionMessage = "Created an InnerException";
-        private const uint COR_E_ENDOFSTREAM = 0x80070026;
-        private const uint E_POINTER = 0x80004003;
-
         [Fact]
-        public static void NullReferenceException_Ctor_Empty()
+        public static void Exception_GetType()
         {
-            NullReferenceException i = new NullReferenceException();
-            Assert.Equal(E_POINTER, (uint)i.HResult);
+            Assert.Equal(typeof(Exception), (new Exception()).GetType());
+            Assert.Equal(typeof(NullReferenceException), (new NullReferenceException()).GetType());
         }
 
         [Fact]
-        public static void NullReferenceException_Ctor_String()
+        public static void Exception_TargetSite()
         {
-            NullReferenceException i = new NullReferenceException(exceptionMessage);
-            Assert.Equal(exceptionMessage, i.Message);
-            Assert.Equal(E_POINTER, (uint)i.HResult);
-        }
+            bool caught = false;
 
-        [Fact]
-        public static void NullReferenceException_Ctor_String_Exception()
-        {
-            Exception ex = new Exception(innerExceptionMessage);
-            NullReferenceException i = new NullReferenceException(exceptionMessage, ex);
+            try
+            {
+                throw new Exception();
+            }
+            catch (Exception ex)
+            {
+                caught = true;
 
-            Assert.Equal(exceptionMessage, i.Message);
-            Assert.Equal(i.InnerException.Message, innerExceptionMessage);
-            Assert.Equal(i.InnerException.HResult, ex.HResult);
-            Assert.Equal(E_POINTER, (uint)i.HResult);
-        }
+                Assert.Equal(MethodInfo.GetCurrentMethod(), ex.TargetSite);
+            }
 
-        [Fact]
-        public static void EndOfStreamException_Ctor_Empty()
-        {
-            EndOfStreamException i = new EndOfStreamException();
-            Assert.Equal(COR_E_ENDOFSTREAM, (uint)i.HResult);
-        }
-
-        [Fact]
-        public static void EndOfStreamException_Ctor_String()
-        {
-            EndOfStreamException i = new EndOfStreamException(exceptionMessage);
-            Assert.Equal(exceptionMessage, i.Message);
-            Assert.Equal(COR_E_ENDOFSTREAM, (uint)i.HResult);
-        }
-
-        [Fact]
-        public static void EndOfStreamException_Ctor_String_Exception()
-        {
-            Exception ex = new Exception(innerExceptionMessage);
-            EndOfStreamException i = new EndOfStreamException(exceptionMessage, ex);
-
-            Assert.Equal(exceptionMessage, i.Message);
-            Assert.Equal(i.InnerException.Message, innerExceptionMessage);
-            Assert.Equal(i.InnerException.HResult, ex.HResult);
-            Assert.Equal(COR_E_ENDOFSTREAM, (uint)i.HResult);
+            Assert.True(caught);
         }
     }
 }
